Improved Temperature Conversion
Created a single function to convert to either degree celsius or
fahrenheit depending on the type (c/F) provided by the user. At the
moment, the functions (main and temperature conversion) do not have
extensive error checking; the program might crash should the user enter
a character or string of characters for start, end and step values which
require floating point precision values.

This is just an improvement of the original code.