First Steps
Created a file to hold the main function with other functions. The
function created so far are push() and constructLinkedList()

 - push() function
This function is passed data of type struct. Each data is added to a new
node with a link. The first node is then made the head.

- constructLinkedList() function
This function of type struct Records is passed an array of data of type
struct and an integer value. The data in the array are added to the node
and the integer value is used to determine the number of nodes to
create. Each node has a link to the next node. A reference to the head is returned
from the function