Fixed Bugs
I have managed to corner the bugs and fixed them up. First, I honestly
reviewed online lessons on structs and practiced a few times, taking
critical note of areas of great concern, which is manipulating array of
characters in a struct and handling complex structs.

I change all pointer variables in the header to arrays with size of at most
30. This is manageable for the beginning. Then, I made several changes in
the main file, running and debugging, critical examining any error message
that came up. This gave me a clearer picture of what was causing the bug.
Now I feel confident to continue my studies in Singly Linked List and then on to Doubly linked list.