Programme improvement
Added more functions and a struct as the lesson progressed

lists.h file
Added a new struct, linked_list with which to create a linked list to be
used to handle collisions that occur in the hash table

hashfunctions.c file
Added three functions to this file:

- allocate_list function:
This function allocates memory space for the variable that invokes it.
It returns a reference to the newly allocated memory area.

- create_new_node function:
This invokes the allocate_list() function for a memory space, creates a
new node and adds the item passed to it as argument to the node. It
points the next address of the node to null and returns a reference to
the new node

- insert_into_linkedlist function:
This function invokes the create_new_node() function and passes the item
parameter to it to create a linked list with the new node. This function
is still in development